# apps/model-gateway-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nova Agent Server",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/admin/model-config/models": {
      "get": {
        "operationId": "listModels",
        "tags": [
          "model-config"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "publicName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 100
                      },
                      "providerId": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                      },
                      "upstreamName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 150
                      },
                      "contextWindow": {
                        "type": "integer",
                        "exclusiveMinimum": true,
                        "maximum": 9007199254740991,
                        "minimum": 0
                      },
                      "maxOutput": {
                        "type": "integer",
                        "exclusiveMinimum": true,
                        "maximum": 9007199254740991,
                        "minimum": 0
                      },
                      "thinkingLevels": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "enum": [
                            "off",
                            "low",
                            "medium",
                            "high",
                            "max"
                          ]
                        }
                      },
                      "parallelToolCalls": {
                        "type": "boolean"
                      },
                      "reasoningFormat": {
                        "type": "string",
                        "enum": [
                          "none",
                          "openai",
                          "anthropic",
                          "deepseek",
                          "minimax"
                        ]
                      },
                      "inputModalities": {
                        "minItems": 1,
                        "type": "array",
                        "items": {
                          "type": "string",
                          "enum": [
                            "text",
                            "image"
                          ]
                        }
                      },
                      "enabled": {
                        "type": "boolean"
                      },
                      "priceIn": {
                        "type": "string",
                        "pattern": "^\\d+(\\.\\d{1,8})?$"
                      },
                      "priceOut": {
                        "type": "string",
                        "pattern": "^\\d+(\\.\\d{1,8})?$"
                      },
                      "priceCacheRead": {
                        "type": "string",
                        "pattern": "^\\d+(\\.\\d{1,8})?$"
                      },
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                      },
                      "providerName": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                      }
                    },
                    "required": [
                      "publicName",
                      "providerId",
                      "upstreamName",
                      "contextWindow",
                      "maxOutput",
                      "thinkingLevels",
                      "parallelToolCalls",
                      "reasoningFormat",
                      "inputModalities",
                      "enabled",
                      "priceIn",
                      "priceOut",
                      "priceCacheRead",
                      "id",
                      "providerName",
                      "createdAt",
                      "updatedAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createModel",
        "tags": [
          "model-config"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "publicName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "providerId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                  },
                  "upstreamName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 150
                  },
                  "contextWindow": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "maximum": 9007199254740991,
                    "minimum": 0
                  },
                  "maxOutput": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "maximum": 9007199254740991,
                    "minimum": 0
                  },
                  "thinkingLevels": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "off",
                        "low",
                        "medium",
                        "high",
                        "max"
                      ]
                    }
                  },
                  "parallelToolCalls": {
                    "type": "boolean"
                  },
                  "reasoningFormat": {
                    "type": "string",
                    "enum": [
                      "none",
                      "openai",
                      "anthropic",
                      "deepseek",
                      "minimax"
                    ]
                  },
                  "inputModalities": {
                    "minItems": 1,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "text",
                        "image"
                      ]
                    }
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "priceIn": {
                    "type": "string",
                    "pattern": "^\\d+(\\.\\d{1,8})?$"
                  },
                  "priceOut": {
                    "type": "string",
                    "pattern": "^\\d+(\\.\\d{1,8})?$"
                  },
                  "priceCacheRead": {
                    "type": "string",
                    "pattern": "^\\d+(\\.\\d{1,8})?$"
                  }
                },
                "required": [
                  "publicName",
                  "providerId",
                  "upstreamName",
                  "contextWindow",
                  "maxOutput",
                  "thinkingLevels",
                  "parallelToolCalls",
                  "reasoningFormat",
                  "inputModalities",
                  "enabled",
                  "priceIn",
                  "priceOut",
                  "priceCacheRead"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "publicName": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "providerId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                    },
                    "upstreamName": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 150
                    },
                    "contextWindow": {
                      "type": "integer",
                      "exclusiveMinimum": true,
                      "maximum": 9007199254740991,
                      "minimum": 0
                    },
                    "maxOutput": {
                      "type": "integer",
                      "exclusiveMinimum": true,
                      "maximum": 9007199254740991,
                      "minimum": 0
                    },
                    "thinkingLevels": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "off",
                          "low",
                          "medium",
                          "high",
                          "max"
                        ]
                      }
                    },
                    "parallelToolCalls": {
                      "type": "boolean"
                    },
                    "reasoningFormat": {
                      "type": "string",
                      "enum": [
                        "none",
                        "openai",
                        "anthropic",
                        "deepseek",
                        "minimax"
                      ]
                    },
                    "inputModalities": {
                      "minItems": 1,
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "text",
                          "image"
                        ]
                      }
                    },
                    "enabled": {
                      "type": "boolean"
                    },
                    "priceIn": {
                      "type": "string",
                      "pattern": "^\\d+(\\.\\d{1,8})?$"
                    },
                    "priceOut": {
                      "type": "string",
                      "pattern": "^\\d+(\\.\\d{1,8})?$"
                    },
                    "priceCacheRead": {
                      "type": "string",
                      "pattern": "^\\d+(\\.\\d{1,8})?$"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                    },
                    "providerName": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    }
                  },
                  "required": [
                    "publicName",
                    "providerId",
                    "upstreamName",
                    "contextWindow",
                    "maxOutput",
                    "thinkingLevels",
                    "parallelToolCalls",
                    "reasoningFormat",
                    "inputModalities",
                    "enabled",
                    "priceIn",
                    "priceOut",
                    "priceCacheRead",
                    "id",
                    "providerName",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/models": {
      "get": {
        "operationId": "listAvailableModels",
        "tags": [
          "models"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                      },
                      "name": {
                        "type": "string"
                      },
                      "providerName": {
                        "type": "string"
                      },
                      "protocol": {
                        "type": "string",
                        "enum": [
                          "openai",
                          "anthropic"
                        ]
                      },
                      "visibility": {
                        "type": "string",
                        "enum": [
                          "public",
                          "owned"
                        ]
                      },
                      "contextWindow": {
                        "type": "integer",
                        "exclusiveMinimum": true,
                        "maximum": 9007199254740991,
                        "minimum": 0
                      },
                      "maxOutput": {
                        "type": "integer",
                        "exclusiveMinimum": true,
                        "maximum": 9007199254740991,
                        "minimum": 0
                      },
                      "reasoningFormat": {
                        "type": "string",
                        "enum": [
                          "none",
                          "openai",
                          "anthropic",
                          "deepseek",
                          "minimax"
                        ]
                      },
                      "thinkingLevels": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "enum": [
                            "off",
                            "low",
                            "medium",
                            "high",
                            "max"
                          ]
                        }
                      },
                      "parallelToolCalls": {
                        "type": "boolean"
                      },
                      "inputModalities": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "enum": [
                            "text",
                            "image"
                          ]
                        }
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "providerName",
                      "protocol",
                      "visibility",
                      "contextWindow",
                      "maxOutput",
                      "reasoningFormat",
                      "thinkingLevels",
                      "parallelToolCalls",
                      "inputModalities"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}